prismaデータテーブル作成＆値の挿入
